pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.34.0",
    "fastapi>=0.128.0",
    "google-generativeai>=0.8.0",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.12",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.9",
    "qdrant-client>=1.16.2",
    "requests>=2.31.0",
    "uvicorn>=0.40.0",
]
